rework dataprocess
