+++ tsPyPackages/tsMqlOverrides/pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "tsmqloverrides"
-version = "2.6.17"
+version = "2.6.18"
 authors = [
   { name = "Tony Shepherd", email = "shep.antony@gmail.com" },
 ]
